+++ src/Manifolds.jl
@@ -48,6 +48,7 @@ import ManifoldsBase:
     check_vector,
     copy,
     copyto!,
+    default_approximation_method,
     default_inverse_retraction_method,
     default_retraction_method,
     default_vector_transport_method,
@@ -149,19 +150,14 @@ import ManifoldsBase:
     submanifold_component,
     submanifold_components,
     vector_space_dimension,
-    vector_transport_along, # just specified in Euclidean - the next 5 as well
-    vector_transport_along_diff,
-    vector_transport_along_project,
+    vector_transport_along,           # just specified in Euclidean - the next 5 as well
     vector_transport_along!,
-    vector_transport_along_diff!,
-    vector_transport_along_project!,
+    vector_transport_along_diff!,     # For consistency these are imported, but for now not
+    vector_transport_along_project!,  # overwritten with new definitons.
     vector_transport_direction,
-    vector_transport_direction_diff,
     vector_transport_direction!,
     vector_transport_direction_diff!,
     vector_transport_to,
-    vector_transport_to_diff,
-    vector_transport_to_project,
     vector_transport_to!,
     vector_transport_to_diff!,
     vector_transport_to_project!, # some overwrite layer 2
@@ -186,6 +182,9 @@ import ManifoldDiff:
     riemannian_Hessian,
     riemannian_Hessian!
 
+import Statistics: mean, mean!, median, median!, cov, var
+import StatsBase: mean_and_var
+
 using Base.Iterators: repeated
 using Distributions
 using Einsum: @einsum
@@ -198,6 +197,7 @@ using ManifoldsBase:
     ℝ,
     ℂ,
     ℍ,
+    AbstractApproximationMethod,
     AbstractBasis,
     AbstractDecoratorManifold,
     AbstractInverseRetractionMethod,
@@ -225,6 +225,7 @@ using ManifoldsBase:
     CotangentSpaceType,
     CoTFVector,
     CoTVector,
+    CyclicProximalPointEstimation,
     DefaultBasis,
     DefaultOrthogonalBasis,
     DefaultOrthonormalBasis,
@@ -232,13 +233,18 @@ using ManifoldsBase:
     DiagonalizingBasisData,
     DiagonalizingOrthonormalBasis,
     DifferentiatedRetractionVectorTransport,
+    EfficientEstimator,
     EmbeddedManifold,
     EmptyTrait,
     EuclideanMetric,
     ExponentialRetraction,
+    ExtrinsicEstimation,
     Fiber,
     FiberType,
     FVector,
+    GeodesicInterpolation,
+    GeodesicInterpolationWithinRadius,
+    GradientDescentEstimation,
     InverseProductRetraction,
     IsIsometricEmbeddedManifold,
     IsEmbeddedManifold,
@@ -295,6 +301,7 @@ using ManifoldsBase:
     VectorSpaceFiber,
     VectorSpaceType,
     VeeOrthogonalBasis,
+    WeiszfeldEstimation,
     @invoke_maker,
     _euclidean_basis_vector,
     combine_allocation_promotion_functions,
@@ -607,8 +614,9 @@ include("deprecated.jl")
 export test_manifold
 export test_group, test_action
 
-#
+# Abstract main types
 export CoTVector, AbstractManifold, AbstractManifoldPoint, TVector
+# Manifolds
 export AbstractSphere, AbstractProjectiveSpace
 export Euclidean,
     ArrayProjectiveSpace,
@@ -748,9 +756,10 @@ export AbstractInverseRetractionMethod,
     ShootingInverseRetraction,
     SoftmaxInverseRetraction
 # Estimation methods for median and mean
-export AbstractEstimationMethod,
+export AbstractApproximationMethod,
     GradientDescentEstimation,
     CyclicProximalPointEstimation,
+    EfficientEstimator,
     GeodesicInterpolation,
     GeodesicInterpolationWithinRadius,
     ExtrinsicEstimation
@@ -788,9 +797,10 @@ export ×,
     convert,
     complex_dot,
     decorated_manifold,
-    default_vector_transport_method,
+    default_approximation_method,
     default_inverse_retraction_method,
     default_retraction_method,
+    default_vector_transport_method,
     det_local_metric,
     differential_canonical_project,
     differential_canonical_project!,