+++ src/manifolds/SymplecticStiefel.jl
@@ -103,10 +103,15 @@ Q_{2n} =
 ````
 The tolerance can be set with `kwargs...` (e.g. `atol = 1.0e-14`).
 """
-function check_point(M::SymplecticStiefel, p; kwargs...)
+function check_point(
+    M::SymplecticStiefel,
+    p::T;
+    atol::Real=sqrt(prod(representation_size(M))) * eps(real(float(number_eltype(T)))),
+    kwargs...,
+) where {T}
     # Perform check that the matrix lives on the real symplectic manifold:
     expected_zero = norm(inv(M, p) * p - I)
-    if !isapprox(expected_zero, 0; kwargs...)
+    if !isapprox(expected_zero, 0; atol=atol, kwargs...)
         return DomainError(
             expected_zero,
             (
@@ -142,14 +147,20 @@ The tolerance can be set with `kwargs...` (e.g. `atol = 1.0e-14`).
 """
 check_vector(::SymplecticStiefel, ::Any...)
 
-function check_vector(M::SymplecticStiefel{<:Any,field}, p, X; kwargs...) where {field}
+function check_vector(
+    M::SymplecticStiefel{S,𝔽},
+    p,
+    X::T;
+    atol::Real=sqrt(prod(representation_size(M))) * eps(real(float(number_eltype(T)))),
+    kwargs...,
+) where {S,T,𝔽}
     n, k = get_parameter(M.size)
     # From Bendokat-Zimmermann: T_pSpSt(2n, 2k) = \{p*H | H^{+} = -H  \}
     H = inv(M, p) * X  # ∈ ℝ^{2k × 2k}, should be Hamiltonian.
-    H_star = inv(Symplectic(2k, field), H)
+    H_star = inv(Symplectic(2k, 𝔽), H)
     hamiltonian_identity_norm = norm(H + H_star)
 
-    if !isapprox(hamiltonian_identity_norm, 0; kwargs...)
+    if !isapprox(hamiltonian_identity_norm, 0; atol=atol, kwargs...)
         return DomainError(
             hamiltonian_identity_norm,
             (